// src/kis/types/financialRatio.type.ts
export type FinancialRatioType = {
  ///결산년원
  stac_yymm: string;
  ///매출액 증가율
  grs: string;
  ///영업이익 증가율
  bsop_prfi_inrt: string;
  ///순이익 증가율
  ntin_inrt: string;
  ///ROE 값
  roe_val: string;
  ///EPS
  eps: string;
  ///주당매출액
  sps: string;
  ///BPS
  bps: string;
  ///유보 비율
  rsrv_rate: string;
  ///부채비율
  lblt_rate: string;
};